Add support for "IDE placeholder" instances of the inbound, asyncreply, and outbound collection elements.  The placeholder instances are added to the model to support the editor but should not be serialized to XML.
The current implementation was to add an isIDEPlaceholder flag to the normal EMF generated impl class, along with an edit listener that on edit of any feature causes the flag to be cleared.  The XML serialization code looks for the flag on save and skips it, on load the XML serialization will check for missing create inbound, asyncreply, and outbound collection elements, and create placeholders instances.

The currently implementation is fairly hacky and could be refactored.