MULEIDE-81 Fix the NPE on save if there is no Services editor
Also added handling of the case where the <model> element is deleted in source.
Current behavior is to remove the Services page if the <model> element is deleted.  The services_diagram file under <project>.settings is deleted.
Support for adding the Services page if the user types in the <model> element in source not implemented.  Closing and reopening the config file *should* display a Services page.  However there is currently a bug where the Services page will fail to be created on reopen due to the model EMF element being set to proxy.