It turns out that one global plugin for common code in Mule IDE and iBeans IDE is not the way to go. While the common code is still a good idea, registering global adapters from the plugin.xml is not - this forces the plugin to become a singleton in the Eclipse runtime. In turn, this would require coordinated releases of both, Mule IDE and iBeans IDE which is not what we want (and what users may want as they would have to update both at the same time, too).
Refactor the code for creating an IdeProject from any selection in the IDE into a standalone class IdeProjectFactory. Without the need for a plugin.xml the whole common ide plugin does not need to be a singleton any more, thus multiple versions of this plugin may be present at runtime. Moreover, each IDE may configure the IdeProjectFactory to its liking now.

Convert MuleIDE to use the new IdeProjectFactory.